Update patch set 1
Patch Set 1:

the other downside to this solution, is the build would now be dependent on the output format of the `readelf -h` command.  I'm not familiar with how often that tool is updated, but if the format ever changed, or the recognized machine formats was ever expanded, the build would break again.

Patch-set: 1
Reviewer: Gerrit User 1076778 <1076778@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0