Update patch set 1
Patch Set 1: Presubmit-Verified+1

=== Finished presubmit run: L76300000158750535 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/653985?revision=1&workplanId=L76300000158750535




TreeHugger finished with: 25 passed.





WARNING: The tip-of-tree(4692674) for git_master/apps may be broken.
The presubmit build passes with the last-known-good-build(4692568), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4692664) for git_master/errorprone_fastbuild3e_linux may be broken.
The presubmit build passes with the last-known-good-build(4692439), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4692674) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4692439), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.








WARNING: The tip-of-tree(4692664) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4692439), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4692664) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4692431), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4692664) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4692431), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4692664) for git_master/test_suites_x86_64_fastbuild3d_linux may be broken.
The presubmit build passes with the last-known-good-build(4692505), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






WARNING: The tip-of-tree(4692664) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4692439), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





Please accept my sincere yet pre-programmed message of undying praise.

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 1
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1