Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Interesting. Just for my own knowledge how does the double $$ syntax work?

It's just escaping the $ so as to defer the work until the inside of the $(eval). $(eval) is basically creating a makefile fragment, and IIRC, we leak those, since the parsers expects most of them to be const, and not worth tracking (from makefiles on disk).

So passing a ~100 character fragment into $(eval) isn't much of a problem. Passing a constantly increasing size into $(eval) uses more and more memory every time.

I suspect the first escape on the line to have the biggest effect, since that's the recursive one. The other ones may help incrementally.

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>