releasetools: Additionally verify against the target build for BBOTA.
While installing a block-based OTA package, we verify the entire
partition against the source build. On verification failure, we call
block_image_verify() to check for feasibility to resume an interrupted
install.

In practice, we found the call to block_image_verify() might fail, if
 - there're multiple partitions to be updated (say /system and /vendor);
 - it's an incremental update that requires stash space on /cache;
 - the interruption happens when updating a later parition (e.g.
   /vendor), with stashed blocks on /cache;
which leads to install failure on resuming an update.

Because block_image_verify() tries to run through the update process for
/system again, without knowing that has been successfully updated. It
tries to allocate the space in need to update /system, but would fail if
the free space on /cache is insufficient (to hold the stashes for both
of /system and /vendor). Note that if the interruption happens during
updating /system, it would have skipped that check automatically based
on the identical stash dirname.

Since this path will only be triggered on resuming, it doesn't affect
the performance for normal update flow.

Bug: 79165963
Test: Generate an incremental package. Check the script.
Test: Simulate the process on device (by triggering a reboot while
      updating /vendor). Check the update log and result.
Change-Id: Ib320fda5355ddf4186c92cd15d0f17561ba8b3d5