Fix escape sequence used in settitle
If I guess correctly, the first escape sequence seems to
be resetting the display attribute. If so, the sequence was "\033]0;",
but this sequence is invalid and doesn't do anything.
The correct sequence should be "\033[0m", note the left bracket
(instead of right), and m (instead of ;).

If I guess incorrectly, this settitle function doesn't work at all,
because the escape seequence is wrong, PROMPT_COMMAND isn't set, and my
old PROMPT_COMMAND is reused (with some buggy behavior).

References:
- http://www.termsys.demon.co.uk/vtansi.htm
- https://en.wikipedia.org/wiki/ANSI_escape_code

Test: set ANDROID_BUILD_SET_WINDOW_TITLE to true, then lunch any target,
see the prompt change
Change-Id: I777c87b4d8e80c0e9437e5d2628f6df020883389