Add a new macro for products making isolation claims.
When a product makes an isolation claim, the build system will verify
that all its artifacts are produced inside its isolation claim, and
fail otherwise.

The claims are all relative to PRODUCT_OUT, and the macro also supports
giving a whitelist, for known existing offenders in the hierarchy. The
build will fail if redundant whitelist entries are present.

Example invocation:
_claims := $(TARGET_COPY_OUT_SYSTEM)/
_whitelist := root/init root/init.rc
$(call make-isolation-claim, $(_claims), $(_whitelist))

Bug: 80410283
Test: Downstream with a new product definition.
Change-Id: I235de681f65254087a42e345af69b2113b682072