Add support enforcing all isolation claims.
Setting PRODUCT_ENFORCE_ISOLATION_CLAIMS to be non-empty will verify
that when an inherited product file makes an isolation claim, no files
other than the ones it produces are allowed inside its claims. This
allows more rigorous control of what goes where, and specifically stops
accidental inclusion of modules in the wrong places (which is very
easy to do otherwise).

In order to enable iterative improvements to current offenders, support
for a whitelist is also added (via the PRODUCT_ISOLATION_CLAIM_WHITELIST
variable). Verification is done that this variable corresponds to
exactly the list of current offenders.

Example use:
  PRODUCT_ENFORCE_ISOLATION_CLAIMS := true
  PRODUCT_ISOLATION_CLAIM_WHITELIST := system/priv-app/Dialer/Dialer.apk

Bug: 80410283
Test: In a downstream CL specifying the above.
Change-Id: I58047db08bde34da21759cfc55f398892b1c809a