[WIP] Invoke $(HIDDENAPI) on all boot dex files simultaneously
The `hiddenapi` build tool encodes information about non-SDK APIs into
boot class path dex files. In the past it would run on each dex file
individually, trying to exact-match dex signatures of its class members
against a provided list of non-SDK APIs. With this approach, `hiddenapi`
can be simple and the heavy lifting of generating the list of non-SDK
APIs is on Doclava.

With Q, we want to move away from Doclava and move the logic to
`hiddenapi` where it can operate directly on dex files rather than
on source files. To this end, `hiddenapi` needs to be have a full view
of the class hierarchy and thus invoked on all dex files at once.

This patch changes the existing definitions to route all boot classpath
dex files through a single build rule.

Test: m
Change-Id: I36d7600ecf1d40127265b35889de7dacaedbe17a