Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Can you just add a small Android.mk in tools/acloud that will install an acloud binary?
> 
>   LOCAL_PATH := $(call my-dir)
> 
>   include $(CLEAR_VARS)
>   LOCAL_MODULE := acloud
>   LOCAL_MODULE_CLASS := EXECUTABLES
>   LOCAL_IS_HOST_MODULE := true
>   LOCAL_SRC_FILES := acloud.sh
>   include $(BUILD_PREBUILT)
> 
> Then `m acloud` will install acloud.sh into out/host/linux-x86/bin/acloud, which is on your PATH. And if you switched over to a python_binary_host in the future, it would look the same, just no longer be a shell script.

Hey DanW, I'd really like it if our Critical User Journey (go/critical-user-journeys from Product Excellence) for the ASuite tools we're building to be simply call env_setup and lunch and the tools are available. This seems to have resonated really well with Atest and makes for a great story. We will still need to build acloud for the infra but any "hard" reasons why we need to extend the user journey with a time-consuming build step before using it if it's just python?

Patch-set: 2
CC: Gerrit User 1083173 <1083173@85c56323-6fa9-3386-8a01-6480fb634889>