Depend on ramdisk.img for BOARD_BUILD_SYSTEM_ROOT_IMAGE
When using BOARD_BUILD_SYSTEM_ROOT_IMAGE, there is no direct dependency
on the files that get typically added to the ramdisk.img apart from the
NOTICE.html dependency that some files have. This change explicitly adds
ramdisk.img as a dependency of system.img so that the image is recreated
any time a file in the ramdisk changes.

Bug: 62387674
Test: ramdisk.img is present in system.img's dependencies
Change-Id: I234cb8d8a648f8963610be519de1b70b7efe1fd7