Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'd rather figure out how to make this automatic.  Otherwise adding a test in a directory that already has a library would require adding this to the library.

Here is the https://cs.corp.google.com/android/art/tools/ahat/Android.mk?dr&g=0&l=33 in question. The situation is that the executable is needed by other tests for execution but also has a test located right next to it.

Any thoughts/initial direction on how to automate this?

One thought I have is that the test config can only be applied to a select white-list of classes, thought I suspect we might need to white-list executables for other use cases...

A second idea is to autogenerate HostTest configs and remove the test xml file that is here already and apply LOCAL_DISABLE_AUTO_GENERATE_TEST_CONFIG to ahat, but that has a similar issue to what you called out for this CL.


Alternatively, a larger fix/refactor, would be to revisit the usage of L_C_S and maybe handle test dependencies in a different manner...

Patch-set: 1