DO NOT SUBMIT: Fall back to linear split upon imgdiff failure (for partners)
Similar to 78205754, we can have omitted zero blocks in the large apk files;
thus leading to the failure when calling imgdiff to split the apk.

Though we have better handling of the condition to use imgdiff in aosp, this cl
serves as an alternative for partners who is using P branch.

Bug: 110962013
Test: generate and verify an ota package for aosp_angler
Change-Id: Ibbe4186f59faa448f255edf869a4d6b4396caf89