Remove dependencies on obj/lib/${name}.so
WIP: Find out why the build rule is missing

This commit removes the dependency from $(LOCAL_BUILT_MODULE) to
`out/host/linux-x86/obj/lib/${name}.so`.  For some unknown reasons,
libc++_host.so is not being installed to
`out/host/linux-x86/obj/lib/${name}.so` and the build rule is not
created for libc++_host.so.  This commit temporary removes such
dependency.

Test: n/a
Change-Id: I349730fc6c6d4037d70a4441acceac777b57ee7e