Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+2
> 
> Verifying API usage seems useful for the apps builds, some apps are only built with TARGET_BUILD_APPS.  We generally check in prebuilts to prebuilts/sdk that the TARGET_BUILD_APPS builds fall back to instead of building from source.  Maybe we should check in the data that veridex needs there?

It actually worked until now (now being me creating the API lists from dex files instead of Doclava). The reason why the new code failed is that there is no 'compile_dex' rule for the prebuilt stubs. I first tried to add that and it worked. But I don't really understand this unbundled build - is it a good or a bad idea to try to build the framework in it?

Other option is prebuilt veridex. We already have that, but it's based on AOSP. We could have an internal version too and use it here. But IIUC, somebody would have to manually keep that updated?

Patch-set: 5
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0