Auto-build list of bundled packages for hidden API
Bundled packages build directly against the framework and they should
be exempted from the hidden API access checks. The information on which
packages should be exempted is communicated to ActivityManagerService
using an XML file with a hardcoded list of package names. This patch
adds a new build flag LOCAL_BUNDLED_MODULE. When set to true, the
associted manifest file is

PREBUILT_INTERNAL.mk

Bug: 79680348
Test: phone boots
Change-Id: Icf37ff70327974f122eb048f5b35baae6c7f29a7