Update patch set 2
Patch Set 2:

When would something compile against platform apis but not need them at runtime?

Global files like this are problematic for the build, because there may be definitions of modules in the build that are not installed on a particular product.  Can it be a per-module file, installed as a dependency of LOCAL_INSTALLED_MODULE?

Patch-set: 2
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>