Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Or be encoded into the AndroidManifest.xml by manifest_fixer.py?

So the reason why they would be different is that right now is that this exemption using an XML file will only work for system apps. For example CTS tests have a different mechanism, but they are still built with PRIVATE_APIS:=true if needed. It would make sense to unify them and a manifest flag ignored for regular apps might be a good way of doing that.

Patch-set: 2
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>