Add core-simple to the standard boot classpath
core-simple.jar is being added to the runtime classpath as
an easy-to-find placeholder and to help identify places that
care about the boot classpath / trigger conversations & issues /
force discussion.

To start, this jar is being added in various locations that know
about core-libart and/or core-oj.  There will be follow-up changes to
make core-simple.jar more than just another core-* jar.

Bug: 113148576
Test: TBD
Change-Id: I267a91b8728f84cfb1b2e8a5c5d157c88b75581c