Update patch set 14
Patch Set 14:

(1 comment)

Hm.  The build errors I expected since that happened the last time too that I tried this.  ccross@ mentioned that system-as-root devices still create a dummy ramdisk just because.  In theory, it'd be good to clean up the dependencies on this dummy ramdisk?  But if it's already the case I can just modify this CL to continue to create one regardless.

The surprising thing is the boot failures, since my local sailfish succeeded.  Perhaps it's a packaging issue.

Patch-set: 14
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>