apidiff: Use Java language version 1.8.
Prior to this CL, apidiff was not passing any -source parameter, which
meant that javadoc was using whatever source level was the default.

Currently, this is language level 9 because the current build toolchain is
OpenJDK 9; this produces warnings because the -bootclasspath parameter
is ignored for language level 9.

This CL fixes this in the simplest possible manner by passing a hard coded
"-source 1.8", which restricts the source language level to OpenJDK 8.
Ideally, the value of JAVA_LANGUAGE_VERSION (java_version in soong) would
be used, but no more than 1.8; in practice, allowing language level 8 even
for lower-level targets should be fine (the lower level will be enforced
through non-docs build targets).

Further, my understanding is that this logic will be removed/replaced in
future, so it doesn't seem worth a more sophisticated fix.

Bug: 112827568
Test: The following command warns "package exists in another module"
      when compiling code from packages in OpenJDK 9's java.base
      (because it doesn't recognize the -bootclasspath argument)
      before but not after this CL:
        make offline-system-sdk-referenceonly-diff

Change-Id: If5151c65c2f92ef86e3108b469a819ac00ac8f85