Add keys for APEX images for eng/userdebug.
These keys will be used to sign and verify APEX modules on
eng/userdebug devices. The keys may still change, but are
added now to unblock apexd development.

The keys were generated as follows:

$ openssl genrsa -out apex_debug_key.pem 4096
$ avbtool extract_public_key --key apex_debug_key.pem -o apex_debug_key

We'll probably need PRODUCT_APEX_KEYS at some point, but we'll wait
with that until we have build support.

Bug: 112684055
Test: /system/etc/security/apex/apex_debug_key found on-device
Change-Id: I5f4b1cb1eb11be6431146aa4297b50205fbc382e