Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Maybe we shouldn't be installing recovery asan binaries, they won't be useful there, recovery doesn't have /data mounted.
> 
> Do you mean not generating *.recovery variant at all when building with asan?

I guess this is fine either way, it just implies we are doing extra useless work.  We would either need to install asan recovery modules to the recovery partition if asan in recovery is desirable, or turn off asan for recovery modules if it is not (for example if it doesn't fit).

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2