Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> (1 comment)
> 
> One of the reasons that I had hidden these warnings (which were overriding commands warnings) was that we didn't really have a good system too be able to provide a "default" for a particular file, but allow devices to override that default. So PRODUCT_COPY_FILES became that.

Makes sense. I can think of ways to enable that.. for example some syntax in PRODUCT_COPY_FILES to allow a particular declaration to be overridden (but the default being not overridable). I can look into that if this need arises. For now, I only want to enable this for a small subset of products.

> 
> I don't think we've gotten any other alternative yet, though maybe we've reduced the need?

I did a spot check of a few internal products that I care about, and they don't appear to have this need (though there are a few bugs that this exposes).

Patch-set: 4
Reviewer: Gerrit User 1060959 <1060959@85c56323-6fa9-3386-8a01-6480fb634889>