Create root mountpoint and symlinks when building images
The root directory requires a set of mountpoints and symlinks that
were previously created using a LOCAL_POST_INSTALL_CMD on the
init and init.environ.rc modules.  If any of these files were removed
there was no way for the build system to know to recreate them.
Move them to the rules that build the ramdisk and system-as-root images
instead.  Recovery image will get similar changes in follow-up CLs.

Bug: 116661769
Test: m checkbuild
Test: Build $OUT/installed-files-root.txt and
      $OUT/installed-files-ramdisk.txt for hikey, taimen and crosshatch
      respectively. Check that there's no change to the installed files.
Change-Id: Ie3febfa3ae65be52f10c5cf3e75879edec48dc48