Update patch set 4
Patch Set 4: Presubmit-Ready+1

I did some experiments with this cl, and both non-A/B and A/B OTA are generated successfully with the python binary. So nothing is actually block us here? The main benefit of the python binary is the directly access of protobuf classes, which is used in the care_map and later transfer commands.

Patch-set: 4
Reviewer: Gerrit User 1088085 <1088085@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1