Update patch set 4
Patch Set 4: Verified-1

This still doesn't work, the rule to build target-files.zip doesn't depend on the rule to build system.img so the directories and symlinks don't get created.  I can't also create the symlinks in the target-files.zip rule because if both rules run simultaneously ln -sf might remove a symlink right as the other rule is trying to zip it.

The only fix I can think of would be to make the target-files.zip rule create the directories and symlinks in some other directory besides $OUT/root and then tell the script to merge that directory into $OUT/root.  It would be likely that the directories and symlinks exist in both $OUT/root and the new directory, but the new directory would be guaranteed to be complete.

Patch-set: 4
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1